fix: validator genesis

+++ state-chain/cf-integration-tests/src/lib.rs
@@ -607,7 +607,7 @@ mod tests {
 					bond: self.accounts[0].1,
 					claim_period_as_percentage: PERCENT_OF_EPOCH_PERIOD_CLAIMABLE,
 					backup_node_percentage: 34,
-					authority_set_min_size: self.min_authorities,
+					authority_set_min_size: self.min_authorities as u8,
 				},
 				ethereum_vault: EthereumVaultConfig {
 					vault_key: ethereum_vault_key,
